+++ cmd/zetae2e/local/local.go
@@ -463,6 +463,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 
 		if !deployerRunner.IsRunningUpgrade() {
 			solanaTests = append(solanaTests, []string{
+				e2etests.TestSolanaDepositThroughProgramName,
 				e2etests.TestSolanaDepositAndCallName,
 				e2etests.TestSolanaWithdrawAndCallName,
 				e2etests.TestZEVMToSolanaCallName,