+++ zetaclient/chains/bitcoin/signer/outbound_data.go
@@ -2,7 +2,6 @@ package signer
 
 import (
 	"fmt"
-	"math"
 	"strconv"
 
 	"github.com/btcsuite/btcd/btcutil"
@@ -29,17 +28,23 @@ type OutboundData struct {
 	amountSats int64
 
 	// feeRate is the fee rate in satoshis/vByte
-	feeRate int64
+	feeRate uint64
 
-	// txSize is the average size of a BTC outbound transaction
-	// user is charged (in ZRC20 contract) at a static txSize on each withdrawal
-	txSize int64
+	// feeRateLatest is the latest median fee rate in satoshis/vByte
+	// this value is fed by the zetacore when it bumps the gas price with gas stability pool
+	feeRateLatest uint64
 
-	// nonce is the nonce of the outbound
-	nonce uint64
+	// feeRateBumpped is a flag to indicate if the fee rate in CCTX is bumped by zetacore
+	feeRateBumped bool
+
+	// minRelayFee is the minimum relay fee in unit of BTC
+	minRelayFee float64
 
 	// height is the ZetaChain block height
 	height uint64
+
+	// nonce is the nonce of the outbound
+	nonce uint64
 
 	// cancelTx is a flag to indicate if this outbound should be cancelled
 	cancelTx bool
@@ -63,9 +68,39 @@ func NewOutboundData(
 	}
 
 	// parse fee rate
-	feeRate, err := strconv.ParseInt(params.GasPrice, 10, 64)
-	if err != nil || feeRate <= 0 {
+	feeRate, err := strconv.ParseUint(params.GasPrice, 10, 64)
+	if err != nil || feeRate == 0 {
 		return nil, fmt.Errorf("invalid fee rate %s", params.GasPrice)
+	}
+
+	// check if zetacore has bumped the fee rate
+	// 'GasPriorityFee' is always "0" for Bitcoin unless zetacore bumps the fee rate
+	var (
+		feeRateBumped bool
+		feeRateLatest uint64
+	)
+	if params.GasPriorityFee != "" {
+		gasPriorityFee, err := strconv.ParseUint(params.GasPriorityFee, 10, 64)
+		if err != nil {
+			return nil, errors.Wrapf(err, "invalid gas priority fee %s", params.GasPriorityFee)
+		}
+
+		if gasPriorityFee > 0 {
+			feeRateBumped = true
+			feeRateLatest = gasPriorityFee
+			logger.Info().Str("latest_fee_rate", params.GasPriorityFee).Msg("fee rate is bumped by zetacore")
+		}
+	}
+
+	// to avoid minRelayTxFee error, please do not use the minimum rate (1 sat/vB by default).
+	// we simply add additional 1 sat/vB to 'minRate' to avoid tx rejection by Bitcoin core.
+	// see: https://github.com/bitcoin/bitcoin/blob/5b8046a6e893b7fad5a93631e6d1e70db31878af/src/policy/policy.h#L42
+	minRate, err := common.FeeRateToSatPerByte(minRelayFee)
+	if err != nil {
+		return nil, errors.Wrapf(err, "invalid min relay fee")
+	}
+	if feeRate <= minRate {
+		feeRate = minRate + 1
 	}
 
 	// check receiver address
@@ -78,22 +113,9 @@ func NewOutboundData(
 	}
 
 	// amount in BTC and satoshis
+	// the float64 'amount' is used later to select UTXOs, precision does not matter
 	amount := float64(params.Amount.Uint64()) / 1e8
 	amountSats := params.Amount.BigInt().Int64()
-
-	// check gas limit
-	if params.CallOptions == nil {
-		// never happens, 'GetCurrentOutboundParam' will create it
-		return nil, errors.New("call options is nil")
-	}
-	if params.CallOptions.GasLimit > math.MaxInt64 {
-		return nil, fmt.Errorf("invalid gas limit %d", params.CallOptions.GasLimit)
-	}
-
-	// add minimum relay fee (1000 satoshis/KB by default) to gasPrice to avoid minRelayTxFee error
-	// see: https://github.com/bitcoin/bitcoin/blob/master/src/policy/policy.h#L35
-	satPerByte := common.FeeRateToSatPerByte(minRelayFee)
-	feeRate += satPerByte
 
 	// compliance check
 	restrictedCCTX := compliance.IsCCTXRestricted(cctx)
@@ -103,9 +125,9 @@ func NewOutboundData(
 	}
 
 	// check dust amount
-	dustAmount := params.Amount.Uint64() < constant.BTCWithdrawalDustAmount
+	dustAmount := amountSats < constant.BTCWithdrawalDustAmount
 	if dustAmount {
-		logger.Warn().Msgf("dust amount %d sats, canceling tx", params.Amount.Uint64())
+		logger.Warn().Int64("amount", amountSats).Msg("outbound will be cancelled due to dust amount")
 	}
 
 	// set the amount to 0 when the tx should be cancelled
@@ -116,14 +138,15 @@ func NewOutboundData(
 	}
 
 	return &OutboundData{
-		to:         to,
-		amount:     amount,
-		amountSats: amountSats,
-		feeRate:    feeRate,
-		// #nosec G115 checked in range
-		txSize:   int64(params.CallOptions.GasLimit),
-		nonce:    params.TssNonce,
-		height:   height,
-		cancelTx: cancelTx,
+		to:            to,
+		amount:        amount,
+		amountSats:    amountSats,
+		feeRate:       feeRate,
+		feeRateLatest: feeRateLatest,
+		feeRateBumped: feeRateBumped,
+		minRelayFee:   minRelayFee,
+		height:        height,
+		nonce:         params.TssNonce,
+		cancelTx:      cancelTx,
 	}, nil
 }